Moving generic Sla discovery code into Junos Class
